Reset Modal forms on close and submit
