+++ www/src/components/account/InviteUser.js
@@ -1,8 +1,13 @@
 import { Button, Div } from 'honorable'
 import {
-  Codeline, MailIcon, Modal, ModalActions, ModalHeader, ValidatedInput,
+  Codeline,
+  MailIcon,
+  Modal,
+  ModalActions,
+  ModalHeader,
+  ValidatedInput,
 } from 'pluralsh-design-system'
-import { useState } from 'react'
+import { useCallback, useState } from 'react'
 
 import { useMutation } from '@apollo/client'
 
@@ -13,11 +18,18 @@ import { GqlError } from '../utils/Alert'
 export function InviteUser() {
   const [open, setOpen] = useState(false)
   const [email, setEmail] = useState('')
-  const [mutation, { loading, data, error }] = useMutation(CREATE_INVITE, {
+  const [invite, setInvite] = useState(null)
+  const resetAndClose = useCallback(() => {
+    setEmail('')
+    setInvite(null)
+    setOpen(false)
+  }, [])
+  const [mutation, { loading, error }] = useMutation(CREATE_INVITE, {
     variables: { attributes: { email } },
+    onCompleted: data => {
+      setInvite(data && data.createInvite)
+    },
   })
-
-  const invite = data && data.createInvite
 
   return (
     <>
@@ -25,20 +37,20 @@ export function InviteUser() {
         <Button
           secondary
           onClick={() => setOpen(true)}
-        >Invite user
+        >
+          Invite user
         </Button>
       </Div>
       <Modal
         open={open}
-        onClose={() => setOpen(false)}
-        width="40vw"
+        onClose={() => resetAndClose()}
+        width="100%"
       >
-        <ModalHeader onClose={() => setOpen(false)}>
-          INVITE USERS
-        </ModalHeader>
+        <ModalHeader>Invite users</ModalHeader>
         <ValidatedInput
+          disabled={!!invite}
           value={email}
-          startIcon={<MailIcon size={15} />}
+          startIcon={<MailIcon />}
           onChange={({ target: { value } }) => setEmail(value)}
           label="Email address"
         />
@@ -50,18 +62,26 @@ export function InviteUser() {
         )}
         {invite && <Codeline marginTop="small">{inviteLink(invite)}</Codeline>}
         <ModalActions>
-          <Button
-            secondary
-            onClick={() => setOpen(false)}
-          >Cancel
-          </Button>
-          <Button
-            onClick={mutation}
-            loading={loading}
-            disabled={email.length === 0}
-            marginLeft="medium"
-          >Invite
-          </Button>
+          {invite ? (
+            <Button onClick={() => resetAndClose()}>Done</Button>
+          ) : (
+            <>
+              <Button
+                secondary
+                onClick={() => resetAndClose()}
+              >
+                Cancel
+              </Button>
+              <Button
+                onClick={mutation}
+                loading={loading}
+                disabled={email.length === 0}
+                marginLeft="medium"
+              >
+                Invite
+              </Button>
+            </>
+          )}
         </ModalActions>
       </Modal>
     </>
